+++ src/main/java/me/ajlane/geo/repast/succession/pathway/io/graph/GraphBasedLcsTransitionMapReader.java
@@ -8,32 +8,32 @@ import me.ajlane.geo.repast.succession.pathway.coded.CodedLcsTransitionMap;
 import me.ajlane.geo.repast.succession.pathway.convert.EnvrStateAliasTranslator;
 import me.ajlane.geo.repast.succession.pathway.convert.LcsTransitionMapConverter;
 
-public class GraphBasedLcsTransitionMapFactory implements LcsTransitionMapFactory {
+public class GraphBasedLcsTransitionMapReader implements LcsTransitionMapFactory {
 
-  final static Logger logger = Logger.getLogger(GraphBasedLcsTransitionMapFactory.class);
+  final static Logger logger = Logger.getLogger(GraphBasedLcsTransitionMapReader.class);
 
   GraphDatabaseService graph;
   String modelID;
   LcsTransitionMapConverter converter;
   AliasedLcsTransitionMap aliasedMap;
 
-  public GraphBasedLcsTransitionMapFactory (GraphDatabaseService graph, String modelID,
+  public GraphBasedLcsTransitionMapReader (GraphDatabaseService graph, String modelID,
       EnvrStateAliasTranslator translator) {
     this.graph = graph;
     this.modelID = modelID;
     this.converter = new LcsTransitionMapConverter(translator);
   }
 
-  GraphBasedLcsTransitionMapFactory (GraphDatabaseService graph,
+  GraphBasedLcsTransitionMapReader (GraphDatabaseService graph,
       EnvrStateAliasTranslator translator) {
     this(graph, null, translator);
   }
 
-  GraphBasedLcsTransitionMapFactory(GraphDatabaseService graph) {
+  GraphBasedLcsTransitionMapReader(GraphDatabaseService graph) {
     this(graph, null, null);
   }
 
-  GraphBasedLcsTransitionMapFactory(GraphDatabaseService graph, String modelID) {
+  GraphBasedLcsTransitionMapReader(GraphDatabaseService graph, String modelID) {
     this(graph, modelID, null);
   }
 
@@ -51,7 +51,7 @@ public class GraphBasedLcsTransitionMapFactory implements LcsTransitionMapFactor
     return converter.convert(getAliasedLcsTransitionMap());
     } catch (NullPointerException e) {
       logger.error("Could not convert aliases to codes as no EnvrStateAliasTranslator "
-                   + "specified in GraphBasedLcsTransitionMapFactory constructor.", e);
+                   + "specified in GraphBasedLcsTransitionMapReader constructor.", e);
       throw e;
     }
   }